+++ src/Foundation/Models/PaymentMethod.php
@@ -14,10 +14,12 @@ declare(strict_types=1);
 
 namespace Vanilo\Foundation\Models;
 
+use Konekt\Address\Concerns\Zoneable;
 use Vanilo\Channel\Traits\Channelable;
 use Vanilo\Payment\Models\PaymentMethod as BasePaymentMethod;
 
 class PaymentMethod extends BasePaymentMethod
 {
     use Channelable;
+    use Zoneable;
 }