+++ Changelog.md
@@ -41,6 +41,7 @@
 - Added the `withImages` and `withChannels` methods to the product search class (eager loads media)
 - Added the `channelables` table for being many-to-many polymorphic relationships with channels and arbitrary models
 - Added the `mapInto()` method to the `RelationAdjustmentCollection` class, which forwards the call to the underlying Eloquent collection
+- Added the `Zoneable` support to the Foundation PaymentMethod class
 - Added the `Channelable` behavior to Foundation Product, MasterProduct, PaymentMethod, ShippingMethod and Taxonomy classes
 - Added and extended Foundation Channel model that contains the known relationships to the "channelable" models
 - Added the `withinChannel[s]` methods to the ProductSearch class